requirements.txt
ansiwrap==0.8.4
appdirs==1.4.3
arrow==0.15.5
asgiref==3.2.7
asteval==0.9.18
attrs==19.3.0
beautifulsoup4==4.8.2
binaryornot==0.4.4
black==19.3b0
boto3==1.7.84
botocore==1.10.84
buckup==0.1a2
certifi==2019.11.28
cffi==1.14.0
cfgv==3.0.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cookiecutter==1.7.0
cryptography==2.8
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.4
django-cogwheels==0.3
django-filter==2.3.0
django-modelcluster==5.0.1
django-storages==1.9.1
django-taggit==1.2.0
django-toolbelt==0.0.1
django-treebeard==4.3.1
djangorestframework==3.11.1
docutils==0.16
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
identify==1.4.11
idna==2.9
jdcal==1.4.1
jeepney==0.4.3
Jinja2==2.11.1
jinja2-time==0.2.0
jmespath==0.9.5
jrnl==2.3
keyring==19.3.0
l18n==2018.5
MarkupSafe==1.1.1
nodeenv==1.3.5
openpyxl==3.0.4
packaging==20.1
parsedatetime==2.5
passlib==1.7.2
pathtools==0.1.2
Pillow==6.2.2
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.1.0
psycopg2==2.8.4
py==1.8.1
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
pyxdg==0.26
PyYAML==5.3
requests==2.23.0
s3transfer==0.1.13
SecretStorage==3.1.2
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
static3==0.7.0
tablib==2.0.0
textwrap3==0.9.2
toml==0.10.0
tox==3.14.4
tzlocal==1.5.1
Unidecode==1.1.1
urllib3==1.25.8
virtualenv==20.0.3
wagtail==2.9.3
wagtailfontawesome==1.2.1
wagtailmenus==3.0.2
watchdog==0.10.2
webencodings==0.5.1
whichcraft==0.6.1
whitenoise==5.0.1
Willow==1.3
xlrd==1.2.0
XlsxWriter==1.3.0
xlwt==1.3.0
